requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.31.2
idna==3.10
inference-sdk==0.49.5
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.39.1
numpy==2.1.3
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
supervision==0.25.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_keras==0.1
termcolor==3.1.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
ultralytics==8.3.137